Fixes case where a `Ref` is treated as a string.
In the build phase basic checking is done on the prefix and suffix
field to determine if they collide with other configuration.

This is only possible if the values for these fields are bound at build time.
Since `Refs` are bound only at apply these checks are not possible at build time.

Ideally these checks would be run at apply time after the refs are bound. I think that
change is beyond the scope of this tactical fix.

Fixes part of: https://github.com/jorgebastida/gordon/issues/34