Lazy generation of usage help
Instead having a string with all the possible actions for gordon, this
string is going to be generated with a lambda after the parser has all
the actions added to it.

Removing the usage kwarg to ArgumentParser we could achieve a similar
output:

  usage: gordon [-h] {startproject,startapp,build,apply,run,delete} ...

But to keep consistency with the previous output the lambda was added.